# contracts/openapi/lumina-gateway-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lumina API Gateway",
    "version": "1.0.0",
    "description": "Public gateway contract for web and external clients."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "paths": {
...
    "/v1/plagiarism/ingest": {
      "post": {
        "summary": "Ingest text into vector storage through gateway",
        "operationId": "gatewayIngest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document ingested.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid JSON or invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid API key when gateway key mode is enabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed upstream.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Upstream unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/plagiarism/check": {
      "post": {
        "summary": "Synchronous plagiarism check",
        "operationId": "gatewayCheck",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Check completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid JSON or invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid API key when gateway key mode is enabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed upstream.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Upstream unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/plagiarism/check-jobs": {
      "post": {
        "summary": "Submit asynchronous plagiarism check job",
        "operationId": "submitCheckJob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckJobAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid JSON or invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid API key when gateway key mode is enabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Job feature unavailable or upstream unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/pdf/extract": {
      "post": {
        "summary": "Extract plain text from uploaded PDF",
        "operationId": "gatewayExtractPdf",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Text extracted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfExtractResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid multipart/PDF payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid API key when gateway key mode is enabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "File too large.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed upstream.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
...
      "IngestRequest": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1
          },
          "document_id": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "IngestResponse": {
        "type": "object",
        "required": [
          "document_id",
          "chunk_count",
          "collection"
        ],
        "properties": {
          "document_id": {
            "type": "string"
          },
          "chunk_count": {
            "type": "integer"
          },
          "collection": {
            "type": "string"
          }
        }
      },
      "CheckRequest": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1
          },
          "top_k": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20,
            "nullable": true
          },
          "threshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          }
        }
      },
      "MatchResult": {
        "type": "object",
        "required": [
          "document_id",
          "chunk_id",
          "matched_text",
          "semantic_score",
          "exact_score",
          "final_score"
        ],
        "properties": {
          "document_id": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "chunk_id": {
            "type": "string"
          },
          "matched_text": {
            "type": "string"
          },
          "semantic_score": {
            "type": "number"
          },
          "exact_score": {
            "type": "number"
          },
          "final_score": {
            "type": "number"
          }
        }
      },
      "CheckResponse": {
        "type": "object",
        "required": [
          "is_plagiarism",
          "overall_score",
          "threshold",
          "matches"
        ],
        "properties": {
          "is_plagiarism": {
            "type": "boolean"
          },
          "overall_score": {
            "type": "number"
          },
          "threshold": {
            "type": "number"
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchResult"
            }
          }
        }
      },
      "CheckJobAcceptedResponse": {
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "properties": {
          "job_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "PdfExtractResponse": {
        "type": "object",
        "required": [
          "filename",
          "page_count",
          "character_count",
          "text"
        ],
        "properties": {
          "filename": {
            "type": "string"
          },
          "page_count": {
            "type": "integer"
          },
          "character_count": {
            "type": "integer"
          },
          "text": {
            "type": "string"
          }
        }
      }
    }
  }
}